requirements.txt
telethon==1.36.0
python-dotenv==1.0.0
httpx==0.25.0
beautifulsoup4==4.12.0
yt-dlp==2024.1.1
